Crate gitjournal

The Git Commit Message and Changelog Generation Framework

This crate contains the library for the git-journal executable. It handles all the parsing and commit message modification stuff which is provided by the executable.

Example usage

use gitjournal::GitJournal;
let mut journal = GitJournal::new(".").unwrap();
journal.parse_log("HEAD", "rc", &1, &false, &true);
journal.print_log(true, None, None).expect("Could not print short log.");

Simply create a new git-journal struct from a given path (. in this example). Then parse the log between a given commit range or a single commit. In this example we want to retrieve everything included in the last git tag, which does not represent a release candidate (contains "rc"). After that parsing the log will be printed in the shortest possible format.

Reexports

pub use config::Config;

Modules

config
Everything related to the git-journal configuration. The configuration files are stored in toml format with the file name .gitjournal.toml.

errors
Basic error handling mechanisms

Structs

GitJournal
The main structure of git-journal.