Struct gitjournal::config::Config

pub struct Config {
    pub categories: Vec<String>,
    pub category_delimiters: Vec<String>,
    pub colored_output: bool,
    pub default_template: Option<String>,
    pub enable_debug: bool,
    pub excluded_commit_tags: Vec<String>,
    pub enable_footers: bool,
    pub show_commit_hash: bool,
    pub show_prefix: bool,
    pub sort_by: String,
    pub template_prefix: String,
}

The configuration structure for git-journal.

Fields
categories: Vec<String>

Specifies the available categories for the commit message

category_delimiters: Vec<String>

Set the characters where the categories are wrapped in

colored_output: bool

Set to false if the output should not be colored

default_template: Option<String>

Specifies the default template. Will be used for tag validation and printing.

enable_debug: bool

Show or hide the debug messages like [OKAY] ... or [INFO] ...

excluded_commit_tags: Vec<String>

Excluded tags in an array, e.g. "internal"

enable_footers: bool

Enable or disable the output and accumulation of commit footers

show_commit_hash: bool

Show or hide the commit hash for every entry

show_prefix: bool

Show or hide the commit message prefix, e.g. JIRA-1234

sort_by: String

Sort the commits during the output by "date" (default) or "name"

template_prefix: String

Commit message template prefix which will be added during commit preparation

Methods

impl Config

fn new() -> Self

Constructs a new Config with default values.

Examples

use gitjournal::Config;
let config = Config::new();

fn save_default_config(&self, path: &str) -> Result<String>

Save the default configuration file in a certain path.

Examples

use gitjournal::Config;
Config::new().save_default_config(".").expect("Could not save config.");

Errors

When toml encoding or file creation failed.

fn load(&mut self, path: &str) -> Result<()>

Load a configuration file from a certain path.

Examples

use gitjournal::Config;
Config::new().load(".").expect("Could not load config.");

Errors

When toml decoding or file opening failed.

fn is_default_config(&self) -> bool

Check if the configuration matches with the default one.

Examples

use gitjournal::Config;
assert_eq!(Config::new().is_default_config(), true);